requirements.txt
requests
pysocks
idna
cryptography
pillow
flask
werkzeug
markupsafe
click
itsdangerous
termcolor
mysql-connector-python